requirements.txt
PyPDF2
streamlit
google-generativeai
IPython
firebase_admin
